Brute force some hacks to make it work?
I'm just going to ignore the realmName check completely for the moment.
Can't figure out why it is coming back blank sometimes.

Found where it is check if they are already in the party - but I've had
reports of people gettign whispers when they are already in so I think
that is broken (spaces in realms maybe?)

Also, not sure if the OnGroupChange event is working right and the
ConvertToRaid() is getting triggered at the correct time?

Suggesting raid leaders invite one person manually, ConvertToRaid via
the UI and then start the mass invites and see how that works.